ports publishing logic completly to wikiPublishButton WIP
