+++ src/components/CreateWiki/CreateWikiTopBar/WikiPublish/WikiPublishButton.tsx
@@ -24,12 +24,16 @@ import { getWikiSlug } from '@/utils/CreateWikiUtils/getWikiSlug'
 import { useWhiteListValidator } from '@/hooks/useWhiteListValidator'
 import { store } from '@/store/store'
 import { postWiki } from '@/services/wikis'
-import { PublishWithCommitMessage } from './WikiPublishWithCommitMessage'
+import { ClientError } from 'graphql-request'
+import { SerializedError } from '@reduxjs/toolkit'
+import { useDispatch } from 'react-redux'
 import OverrideExistingWikiDialog from '../../EditorModals/OverrideExistingWikiDialog'
+import { PublishWithCommitMessage } from './WikiPublishWithCommitMessage'
 
 export const WikiPublishButton = () => {
   const toast = useToast()
   const wiki = useAppSelector(state => state.wiki)
+  const dispatch = useDispatch()
   const [submittingWiki, setSubmittingWiki] = useBoolean()
   const { address: userAddress, isConnected: isUserConnected } = useAccount()
   const { userCanEdit } = useWhiteListValidator(userAddress)
@@ -44,6 +48,69 @@ export const WikiPublishButton = () => {
 
   const isPublishDisabled = submittingWiki || !userCanEdit
   const isNewWiki = false
+  const { saveHashInTheBlockchain, signing, verifyTrxHash } = useGetSignedHash()
+
+  useEffect(() => {
+    async function verifyTransactionHash() {
+      if (txHash) verifyTrxHash()
+    }
+    verifyTransactionHash()
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [txHash, verifyTrxHash])
+
+  const processWikiPublishError = (wikiResult: {
+    error: Pick<ClientError, 'name' | 'message' | 'stack'> | SerializedError
+  }) => {
+    setIsLoading('error')
+    let logReason = 'NO_IPFS'
+    if (wikiResult && 'error' in wikiResult) {
+      const rawErrMsg = wikiResult.error.message
+      const prefix = 'Http Exception:'
+      if (rawErrMsg?.startsWith(prefix)) {
+        const errObjString = rawErrMsg.substring(prefix.length)
+        const errObj = JSON.parse(errObjString)
+        // eslint-disable-next-line no-console
+        console.error({ ...errObj })
+        const wikiError =
+          errObj.response.errors[0].extensions.exception.response
+        logReason = wikiError.error
+        setMsg(ValidationErrorMessage(logReason))
+      } else {
+        setMsg(defaultErrorMessage)
+      }
+    }
+    logEvent({
+      action: 'SUBMIT_WIKI_ERROR',
+      label: await getWikiSlug(wiki),
+      category: 'wiki_error',
+      value: 1,
+    })
+  }
+
+  const updateCommitMessage = (override: boolean) => {
+    let wikiCommitMessage = getWikiMetadataById(
+      wiki,
+      EditSpecificMetaIds.COMMIT_MESSAGE,
+    )?.value
+
+    if (isNewWiki) {
+      if (override) {
+        wikiCommitMessage = 'Wiki Overridden 🔄'
+      } else if (revision) {
+        wikiCommitMessage = `Reverted to revision ${revision} ⏪`
+      } else {
+        wikiCommitMessage = 'New Wiki Created 🎉'
+      }
+    }
+
+    dispatch({
+      type: 'wiki/updateMetadata',
+      payload: {
+        id: EditSpecificMetaIds.COMMIT_MESSAGE,
+        value: wikiCommitMessage,
+      },
+    })
+  }
 
   const handleWikiPublish = async (override?: boolean) => {
     if (!isValidWiki(toast, wiki)) return
@@ -55,45 +122,28 @@ export const WikiPublishButton = () => {
       value: 1,
     })
 
-    let wikiCommitMessage = getWikiMetadataById(
-      wiki,
-      EditSpecificMetaIds.COMMIT_MESSAGE,
-    )?.value
-
     if (isUserConnected && userAddress) {
-      if (
+      const ifWikiExists =
         isNewWiki &&
         !override &&
         (await isWikiExists(await getWikiSlug(wiki), setExistingWikiData))
-      ) {
+
+      if (ifWikiExists) {
         onOverrideModalOpen()
         return
       }
 
-      if (isNewWiki) {
-        if (override) {
-          wikiCommitMessage = 'Wiki Overridden 🔄'
-        } else if (revision) {
-          wikiCommitMessage = `Reverted to revision ${revision} ⏪`
-        } else {
-          wikiCommitMessage = 'New Wiki Created 🎉'
-        }
-      }
+      updateCommitMessage(!!override)
 
       setOpenTxDetailsDialog(true)
+
       setSubmittingWiki.on()
 
       const finalWiki = {
         ...wiki,
         user: { id: userAddress },
         content: sanitizeContentToPublish(String(wiki.content)),
-        category: 'daos',
-        metadata: [
-          ...wiki.metadata.filter(
-            m => m.id !== EditSpecificMetaIds.COMMIT_MESSAGE,
-          ),
-          { id: EditSpecificMetaIds.COMMIT_MESSAGE, value: wikiCommitMessage },
-        ].filter(m => m.value),
+        metadata: wiki.metadata.filter(m => m.value),
       }
 
       if (finalWiki.id === CreateNewWikiSlug)
@@ -110,31 +160,7 @@ export const WikiPublishButton = () => {
       if (wikiResult && 'data' in wikiResult) {
         saveHashInTheBlockchain(String(wikiResult.data))
       } else {
-        setIsLoading('error')
-        let logReason = 'NO_IPFS'
-        // get error message from wikiResult
-        if (wikiResult && 'error' in wikiResult) {
-          const rawErrMsg = wikiResult.error.message
-          const prefix = 'Http Exception:'
-          if (rawErrMsg?.startsWith(prefix)) {
-            const errObjString = rawErrMsg.substring(prefix.length)
-            const errObj = JSON.parse(errObjString)
-            // eslint-disable-next-line no-console
-            console.error({ ...errObj })
-            const wikiError =
-              errObj.response.errors[0].extensions.exception.response
-            logReason = wikiError.error
-            setMsg(ValidationErrorMessage(logReason))
-          } else {
-            setMsg(defaultErrorMessage)
-          }
-        }
-        logEvent({
-          action: 'SUBMIT_WIKI_ERROR',
-          label: await getWikiSlug(wiki),
-          category: 'wiki_error',
-          value: 1,
-        })
+        processWikiPublishError(wikiResult)
       }
 
       setSubmittingWiki.off()
@@ -144,7 +170,7 @@ export const WikiPublishButton = () => {
   return (
     <>
       <Tooltip
-        isDisabled={userCanEdit}
+        isDisabled={!!userCanEdit}
         p={2}
         rounded="md"
         placement="bottom-start"
